Move help blueprint to structure module
